backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6